backend/requirements.txt
fastapi
uvicorn
pydantic
tensorflow
scikit-learn
python-chess
pandas
numpy
joblib
python-chess
tqdm